// src/app/enums/enumOrderStatusType.ts
export enum EnumOrderStatusType {
    RecienPedido = "5a59d9853190390014228219",
    Pendiente = "5a59d9853190390014228220",
    ParaEntregar = "5a59d9853190390014228221",
    Entregada = "5a59d9853190390014228222",
    Cancelada = "5a59d9853190390014228224",
    Borrado = "5a59d9853190390014228225"
}